fix(game): wave-6 round-2 — NPC spawn-path bugs, RNG hoist/draw-order batch, complete NPC_AI_Sniper port, VectorMA macro-semantics fixes
NPC_spawn: npc-kill TeamTable crash, MASK_SOLID spawn traces, npc-score
FOFS(targetname), default-falls-into-WP_BLASTER switch, goto-finish
tempGoal-fail return, FRAMETIME. RNG stream: five hoisted-draw reorders
(ST_CheckMoveState + its missing return, Rancor/Wampa stale-legsTimer
reads, Mark1, Grenadier), G_DroidSounds fall-through TIMER_Set for
unmatched classes, NPC_HandleAIFlags arg order verified empirically
against g++-16 (left-to-right). NPC_BSWander tail un-nested;
Jedi_CheckDanger §19 guard (oracle reads alertEvents[-1], Rust would
panic). Vehicle NPCs: FighterNPC restores the live G_DamageFromKiller
landing-death block + 6 sin(serverTime*0.001) f64 chains; AnimalNPC
drops the #ifndef _JK2MP auto-aim block (dead in MP); SpeederNPC ports
the live electrify yaw-wobble. NPC_AI_Sniper completed 15/15 — all
silent stubs replaced with faithful bodies (restored traces, G_SetEnemy,
aim math; YAW/PITCH indices and SPF_NO_HIDE corrected).

VectorMA/VectorScale are the live #if 1 MACROS (q_shared.h:1352) — an
RNG call in the scale expression draws once PER COMPONENT. Three sites
fixed to per-component draws (Sniper miss-loop, Jedi aim-lead); the
Seeker 'fix' built on the dead _VectorMA signature was reverted. Caught
in gate-1 diff review + adversarial validation. Full referee suite
(mock + real-map duel1) green.

+++ crates/mp/game/src/NPC_AI_Remote.rs
@@ -1,4 +1,4 @@
-// PORT-COMPLETE: NPC_AI_Remote.c 1/10
+// PORT-COMPLETE: NPC_AI_Remote.c
 //! FAITHFUL port of `oracle/codemp/game/NPC_AI_Remote.c`.
 //!
 //! One function ported; ten parked due to ambient-state infrastructure.
@@ -291,6 +291,8 @@ pub fn Remote_Hunt(ctx: &mut GameContext, visible: qboolean, advance: qboolean, 
         }
     } else {
         unsafe {
+            // §19: oracle derefs `NPC->enemy->r.currentOrigin` unconditionally.
+            // Source: oracle/codemp/game/NPC_AI_Remote.c:211
             if let Some(enemy_id) = (*npc).enemy {
                 let enemy_ent = &mut ctx.world.g_entities[enemy_id.0 as usize] as *mut gentity_t;
                 crate::q_math::_VectorSubtract(
@@ -470,12 +472,15 @@ pub fn Remote_Attack(ctx: &mut GameContext) {
     }
 
     idealDist = MIN_DISTANCE_SQR + (MIN_DISTANCE_SQR * ctx.world.bg_state.rng.flrand(0.0, 1.0));
-    advance = if distance > idealDist * 1.25 {
+    // C: `distance`(int) compared against `idealDist`(float)*1.25/0.75(double);
+    // the products promote to double and the compare is in double.
+    // Source: oracle/codemp/game/NPC_AI_Remote.c:317-318
+    advance = if (distance as f64) > idealDist as f64 * 1.25 {
         qtrue
     } else {
         qfalse
     };
-    retreat = if distance < idealDist * 0.75 {
+    retreat = if (distance as f64) < idealDist as f64 * 0.75 {
         qtrue
     } else {
         qfalse
